+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 26/05
 - basic mob movement
 - add room mob reset timer (5 min)
+- update mob sprite
+- add rooms to overworld, add mobs to rooms
+- add slash state for player (sword use)
 
 18/05
 - rerwite overworld rendering